Commiting the changed structure, tested code and newly added test code.
