+++ modules/webapp-management/org.wso2.appserver.webapp.security/src/test/resources/testng.xml
@@ -1,0 +1,27 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!--
+Copyright (c) 2015, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+-->
+
+<!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd" >
+
+<suite name="WSO2_App_Server_Test-Suite">
+    <test name="app-server-unit-tests" preserve-order="true" parallel="false">
+        <classes>
+            <class name="org.wso2.appserver.webapp.security.sso.SSOUtilsTest"/>
+            <class name="org.wso2.appserver.webapp.security.sso.ModelClassTest"/>
+        </classes>
+    </test>
+</suite>
